Fix Read the Docs builds
Read the Docs builds began permanently failing on February 26, 2024.

The differential diagnosis between success and failure
seems to be that the last successful build used Poetry 1.7.1;
the first failing build used Poetry 1.8.1.
For reference, Poetry 1.8.1 was released on February 26, 2024.

Read the Docs later updated its "Install dependencies with Poetry"
documentation to note that Poetry needs a `$VIRTUAL_ENV`
environment variable set.

https://docs.readthedocs.com/platform/stable/build-customization.html#install-dependencies-with-poetry

This change aligns the Read the Docs config with the documented
fix, and simultaneously updates the build OS and Python version,
which are both very old.