Migrate to PEP 621 project metadata
These changes now allow the `poetry check` command
to run successfully and without warnings.

As a part of this change, the `docs` extra was removed
in favor of the Poetry `docs` dependency group.

The built sdist was checked and confirmed to still have
the `tests/` directory, as well as the `py.typed` file.
The built wheel was checked and confirmed to still have
the `py.typed` file.