Move the `docs/contributing.rst` contents to `CONTRIBUTING.rst`
This change makes it easier for people to contribute
to the project when opening a PR.

Currently, GitHub shows links to `CONTRIBUTING.rst`
when opening a PR, but when people click on those links
they are simply told "Go to a third-party website
to view the contributing guidelines".

This change replaces the manual redirect with actual content
so that no third-party website (in this case, Read the Docs)
is necessary to understand the contributing guidelines.

In addition, `docs/contributing.rst` now uses a directive
to `include` the contents of `CONTRIBUTING.rst`.
This was tested by building the documentation locally
and confirming that the HTML documentation still contained
the contributing guidelines.

The result is a streamlined contributing experience from within
GitHub, no duplicated content, and no changes to the
built documentation.